Mermaid.jl

Mermaid.Mermaid — Module

Mermaid.jl

Mermaid.jl is a hybrid and multiscale simulation environment in Julia.

Complex simulations can be produced by connecting together components from a wide range of Julia modeling tools.

Its key features are:

It is particularly well suited towards hybrid (continuous and discrete time), multiscale and nested systems. With direct support for nesting models within other models (Agent-based models where each agent solves an ODE, for example).

Models can be specified as arbitrary Julia code, include calls to other programming languages such as C or Python.

Out-of-the-box support for Agents.jl, DifferentialEquations.jl (and related packages), Surrogates.jl, MethodOfLines.jl, and TrixiParticles.jl.

A simple integrator interface allows user-defined extensions to previously unsupported components.

Getting started

Mermaid can be installed from Julia with:

using Pkg; Pkg.add("Mermaid")

We have a begginer friendly tutorial, and guided examples for different features in Mermaid available in the docs.

We also have a few more complex, unguided examples in the examples directory.